refactor: add methods from services in window view model
