+++ Agendai/Services/views/WeekViewService.cs
@@ -45,5 +45,25 @@ namespace Agendai.Services.Views
                 rows.Add(row);
             }
         }
+        
+        public static (int WeekNumber, DateTime StartOfWeek, DateTime EndOfWeek) GetWeekOfMonthRange(DateTime date)
+        {
+            var firstDayOfMonth = new DateTime(date.Year, date.Month, 1);
+            var dayOfWeekOffset = (int)firstDayOfMonth.DayOfWeek; 
+            
+            int totalDays = (date - firstDayOfMonth).Days + dayOfWeekOffset;
+            int weekNumber = (int)Math.Floor(totalDays / 7.0) + 1;
+            
+            var startOfWeek = date.Date.AddDays(-(int)date.DayOfWeek);
+
+            if (startOfWeek.Month < date.Month)
+                startOfWeek = new DateTime(date.Year, date.Month, 1);
+
+            var endOfWeek = startOfWeek.AddDays(6);
+            if (endOfWeek.Month > date.Month)
+                endOfWeek = new DateTime(date.Year, date.Month, DateTime.DaysInMonth(date.Year, date.Month));
+
+            return (weekNumber, startOfWeek, endOfWeek);
+        }
     }
 }
